refactor: split terminal startup helpers
