+++ rust/alera-cli/src/terminal_host/server.rs
@@ -117,6 +117,7 @@ mod terminal_input_requests;
 mod terminal_launch_defaults;
 mod terminal_session_requests;
 mod terminal_spawn;
+mod terminal_startup_commands;
 mod workspace_pinning;
 mod workspace_sidebar_requests;
